fix(Editor): prevent crash on build due to UnityEditor
The method used for generating the labels in the Unity Custom Editors
was located in the `Utilities` script which is included in the main
build, however this script included the `UnityEditor` package which
doesn't get included at build time and therefore would crash any
builds.

The solution is to have an Editor Utilities script within the Editor
directory as that doesn't get included in the build and keeps the
build files clean.
