+++ Assets/VRTK/Editor/VRTK_EditorUtilities.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: cb016f73234cb024baba2ba9330a83d0
+timeCreated: 1474353707
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
