+++ Assets/VRTK/Scripts/Helper/Utilities.cs
@@ -1,8 +1,6 @@
 ﻿namespace VRTK
 {
     using UnityEngine;
-    using UnityEditor;
-    using System;
     using System.Reflection;
     using Highlighters;
 
@@ -159,14 +157,6 @@
             }
 
             return objectHighlighter;
-        }
-
-        public static GUIContent BuildGUIContent<T>(string fieldName, string displayOverride = null)
-        {
-            var displayName = (displayOverride != null ? displayOverride : ObjectNames.NicifyVariableName(fieldName));
-            var fieldInfo = typeof(T).GetField(fieldName);
-            var tooltipAttribute = (TooltipAttribute)Attribute.GetCustomAttribute(fieldInfo, typeof(TooltipAttribute));
-            return (tooltipAttribute == null ? new GUIContent(displayName) : new GUIContent(displayName, tooltipAttribute.tooltip));
         }
     }
 }
